# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "Claim": {
        "properties": {
          "classification": {
            "$ref": "#/components/schemas/Classification"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "end_ms": {
            "minimum": 0.0,
            "title": "End Ms",
            "type": "integer"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "exact_text": {
            "minLength": 1,
            "title": "Exact Text",
            "type": "string"
          },
          "fixture_mode": {
            "default": true,
            "title": "Fixture Mode",
            "type": "boolean"
          },
          "normalized_text": {
            "minLength": 1,
            "title": "Normalized Text",
            "type": "string"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "speaker_label": {
            "title": "Speaker Label",
            "type": "string"
          },
          "start_ms": {
            "minimum": 0.0,
            "title": "Start Ms",
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/ClaimState"
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Verdict"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "public_id",
          "session_id",
          "speaker_label",
          "exact_text",
          "normalized_text",
          "start_ms",
          "end_ms",
          "classification",
          "state",
          "created_at"
        ],
        "title": "Claim",
        "type": "object"
      },
      "ClaimState": {
        "enum": [
          "CAPTURING",
          "TRANSCRIBING",
          "CLAIM_CANDIDATE",
          "CHECKING",
          "EVIDENCE_READY",
          "SYNTHESIZING",
          "COMPLETE",
          "INSUFFICIENT_EVIDENCE",
          "FAILED"
        ],
        "title": "ClaimState",
        "type": "string"
      },
      "Classification": {
        "enum": [
          "opinion",
          "factual_claim",
          "unverifiable"
        ],
        "title": "Classification",
        "type": "string"
      },
      "Evidence": {
        "properties": {
          "canonical_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Canonical Url",
            "type": "string"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "excerpt": {
            "title": "Excerpt",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "independent_key": {
            "default": "legacy",
            "maxLength": 200,
            "minLength": 1,
            "title": "Independent Key",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "publisher": {
            "title": "Publisher",
            "type": "string"
          },
          "query_role": {
            "$ref": "#/components/schemas/SearchRole",
            "default": "neutral"
          },
          "retrieved_at": {
            "format": "date-time",
            "title": "Retrieved At",
            "type": "string"
          },
          "source_tier": {
            "$ref": "#/components/schemas/SourceTier"
          },
          "stance": {
            "$ref": "#/components/schemas/Stance"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "stance",
          "title",
          "canonical_url",
          "publisher",
          "retrieved_at",
          "excerpt",
          "source_tier",
          "content_hash"
        ],
        "title": "Evidence",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Pairing": {
        "properties": {
          "code": {
            "default": "123456",
            "title": "Code",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          }
        },
        "required": [
          "expires_at"
        ],
        "title": "Pairing",
        "type": "object"
      },
      "PushSubscription": {
        "properties": {
          "device_id": {
            "title": "Device Id",
            "type": "string"
          },
          "endpoint": {
            "default": "fake://local",
            "title": "Endpoint",
            "type": "string"
          }
        },
        "required": [
          "device_id"
        ],
        "title": "PushSubscription",
        "type": "object"
      },
      "SearchRole": {
        "enum": [
          "neutral",
          "support",
          "counter"
        ],
        "title": "SearchRole",
        "type": "string"
      },
      "SessionCreate": {
        "properties": {
          "fixture_mode": {
            "default": true,
            "title": "Fixture Mode",
            "type": "boolean"
          },
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "video_title": {
            "default": "Hero demo",
            "title": "Video Title",
            "type": "string"
          },
          "video_url": {
            "default": "https://youtube.com/watch?v=hero",
            "title": "Video Url",
            "type": "string"
          }
        },
        "required": [
          "idempotency_key"
        ],
        "title": "SessionCreate",
        "type": "object"
      },
      "SessionCreated": {
        "properties": {
          "credential": {
            "title": "Credential",
            "type": "string"
          },
          "fixture_mode": {
            "default": true,
            "title": "Fixture Mode",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "credential"
        ],
        "title": "SessionCreated",
        "type": "object"
      },
      "SourceTier": {
        "enum": [
          "primary",
          "research",
          "established",
          "other"
        ],
        "title": "SourceTier",
        "type": "string"
      },
      "Stance": {
        "enum": [
          "support",
          "counter",
          "context"
        ],
        "title": "Stance",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Verdict": {
        "properties": {
          "citation_ids": {
            "items": {
              "type": "string"
            },
            "title": "Citation Ids",
            "type": "array"
          },
          "common_ground": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Common Ground"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "counterevidence_summary": {
            "title": "Counterevidence Summary",
            "type": "string"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "label": {
            "$ref": "#/components/schemas/VerdictLabel"
          },
          "model_name": {
            "default": "hero-fixture",
            "title": "Model Name",
            "type": "string"
          },
          "model_provider": {
            "default": "fake",
            "title": "Model Provider",
            "type": "string"
          },
          "prompt_version": {
            "default": "phase1",
            "title": "Prompt Version",
            "type": "string"
          },
          "uncertainty": {
            "title": "Uncertainty",
            "type": "string"
          }
        },
        "required": [
          "label",
          "confidence",
          "explanation",
          "uncertainty",
          "counterevidence_summary",
          "citation_ids"
        ],
        "title": "Verdict",
        "type": "object"
      },
      "VerdictDraft": {
        "additionalProperties": false,
        "properties": {
          "citation_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 3,
            "minItems": 1,
            "title": "Citation Ids",
            "type": "array"
          },
          "claim_public_id": {
            "title": "Claim Public Id",
            "type": "string"
          },
          "common_ground": {
            "anyOf": [
              {
                "maxLength": 400,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Common Ground"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "counterevidence_summary": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Counterevidence Summary",
            "type": "string"
          },
          "explanation": {
            "maxLength": 700,
            "minLength": 1,
            "title": "Explanation",
            "type": "string"
          },
          "label": {
            "$ref": "#/components/schemas/VerdictLabel"
          },
          "model_name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Model Name",
            "type": "string"
          },
          "model_provider": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Model Provider",
            "type": "string"
          },
          "prompt_version": {
            "const": "phase3-v1",
            "default": "phase3-v1",
            "title": "Prompt Version",
            "type": "string"
          },
          "uncertainty": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Uncertainty",
            "type": "string"
          }
        },
        "required": [
          "claim_public_id",
          "label",
          "confidence",
          "explanation",
          "uncertainty",
          "counterevidence_summary",
          "citation_ids",
          "model_provider",
          "model_name"
        ],
        "title": "VerdictDraft",
        "type": "object"
      },
      "VerdictLabel": {
        "enum": [
          "Supported",
          "Misleading",
          "Disputed",
          "Unsupported",
          "Insufficient evidence"
        ],
        "title": "VerdictLabel",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "Verity API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "health_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/readyz": {
      "get": {
        "operationId": "ready_readyz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Ready"
      }
    },
    "/v1/claims/{public_id}": {
      "get": {
        "operationId": "claim_v1_claims__public_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "public_id",
            "required": true,
            "schema": {
              "title": "Public Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Claim"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim"
      }
    },
    "/v1/claims/{public_id}/verdict": {
      "post": {
        "operationId": "verdict_v1_claims__public_id__verdict_post",
        "parameters": [
          {
            "in": "path",
            "name": "public_id",
            "required": true,
            "schema": {
              "title": "Public Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerdictDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Claim"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verdict"
      }
    },
    "/v1/pairings": {
      "post": {
        "operationId": "pairing_v1_pairings_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pairing"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Pairing"
      }
    },
    "/v1/push-subscriptions": {
      "post": {
        "operationId": "subscribe_v1_push_subscriptions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscription"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Subscribe"
      }
    },
    "/v1/push-subscriptions/{subscription_id}": {
      "delete": {
        "operationId": "unsubscribe_v1_push_subscriptions__subscription_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "subscription_id",
            "required": true,
            "schema": {
              "title": "Subscription Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unsubscribe"
      }
    },
    "/v1/sessions": {
      "post": {
        "operationId": "session_v1_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Session"
      }
    },
    "/v1/sessions/{session_id}/claims": {
      "post": {
        "operationId": "hero_v1_sessions__session_id__claims_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Claim"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Hero"
      }
    }
  }
}